Query pagination previous and next: change the markup rendered on the first and last pages.
The previous  and next page components will render a <span> instead of an <a> on first and last pages of the pagination.

The Previous component will render a <span> instead of an <a> if we are navigating the first page. The Next component will render a <span> instead of an <a> if we are navigating the last page.